refactor: split analytics controller into overview + trends + forecast (#2253)
